requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
spacy
scikit-learn
PyPDF2
python-docx
reportlab